+++ tests/Common/config_pretty_print_true.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symfony\Component\DependencyInjection\Loader\Configurator;
+
+return static function (ContainerConfigurator $container): void {
+    $container->extension('mitopp_schema_org', [
+        'pretty_print' => true,
+    ]);
+};